requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pillow==8.3.2
opencv-python==4.5.3
numpy==1.21.2
pandas==1.3.3
requests==2.26.0
beautifulsoup4==4.10.0
jinja2==3.0.2
python-multipart==0.0.5
lxml==4.6.3
